+++ pkg/pb/v1/series_test.go
@@ -92,6 +92,14 @@ func TestMarshalAndUnmarshalSeries(t *testing.T) {
 				{Value: &modelv1.TagValue_Int{Int: &modelv1.Int{Value: 123}}},
 			}},
 		},
+		{
+			name: "topn",
+			src: &Series{Subject: "subject", EntityValues: []*modelv1.TagValue{
+				{Value: &modelv1.TagValue_Str{Str: &modelv1.Str{Value: "stringValue"}}},
+				{Value: &modelv1.TagValue_Int{Int: &modelv1.Int{Value: 1}}},
+				{Value: &modelv1.TagValue_Int{Int: &modelv1.Int{Value: 46}}},
+			}},
+		},
 	}
 
 	for _, tt := range tests {