+++ CHANGES.md
@@ -2,11 +2,14 @@
 
 Release Notes.
 
-## 0.7.0
+## 0.6.1
 
 - Add benchmarks for stream filtering and sorting.
 
 ### Features
+
+- Limit the max pre-calculation result flush interval to 1 minute.
+- Use both datapoint timestamp and server time to trigger the flush of topN pre-calculation result.
 
 ### Bugs
 
@@ -14,6 +17,8 @@ Release Notes.
 - Data race in the hot series index selection.
 - Remove SetSchema from measure cache which could change the schema in the cache.
 - Fix duplicated items in the query aggregation top-n list.
+- Fix non-"value" field in topN pre-calculation result measure is lack of data.
+- Encode escaped characters to int64 bytes to fix the malformed data.
 
 ## 0.6.0